refactor: split SyncManager.swift into 3 files — zero SwiftLint violations
SyncManager.swift was 2099 lines with 12 SwiftLint violations:
- 2x cyclomatic_complexity (functions with complexity 21-22, limit 15)
- 1x file_length (2099 lines, limit 1700)
- 9x line_length (>120 chars)

Changes:
1. Extract SyncTypes.swift (176 lines) — all type declarations:
   - EventData, PullResult, ProjectionResponse, projection models
   - WebSocket stream message types (SyncStreamRequest, etc.)
   - SyncError and ConnectionStatus enums

2. Extract SyncManager+ProjectionSync.swift (421 lines) — projection sync:
   - syncViaProjections(), fetchProjectionPage(), populateFromProjections()
   - Break populateFromProjections (complexity 21) into 5 sub-functions:
     populateArcs, populateTags, populateStacks, populateTasks, populateReminders
   - Move nonisolated helpers: parseISO8601, dateFromUnixMs, parseStackStatus, parseTaskStatus

3. Refactor streamEventsViaWebSocket (complexity 22) into 5 sub-functions:
   connectStreamWebSocket, sendStreamRequest, handleStreamStart,
   handleStreamBatch, handleStreamComplete, handleStreamError

4. Access level changes for extension compatibility:
   - session → syncSession (internal)
   - modelContainer → syncModelContainer (internal)
   - _isInitialSyncInProgress → isInitialSyncActive (internal)
   - _initialSyncEventsProcessed → syncEventsProcessed (internal)
   - _initialSyncTotalEvents → syncTotalEvents (internal)
   - refreshToken(), saveLastSyncCheckpoint(), generateSyncId() → internal
   - iso8601Standard, iso8601WithFractionalSeconds → internal static

Result: SyncManager.swift 2099→1619 lines, 12→0 SwiftLint violations
Total across all 3 files: 2216 lines (net +117 from sub-function signatures)
